fix: Change maxDecompresedSize to int64 (#20488)

+++ pkg/distributor/distributor_test.go
@@ -2656,7 +2656,7 @@ func TestConfig_Validate(t *testing.T) {
 	tests := []struct {
 		name                        string
 		cfg                         Config
-		expectedMaxDecompressedSize int
+		expectedMaxDecompressedSize int64
 		expectedError               string
 	}{
 		{
